Added extension methods on gdal::RasterBand.
