+++ Cargo.toml
@@ -30,7 +30,7 @@ default = ["masked", "serde"]
 masked = []
 # Enables serde
 serde = ["dep:serde"]
-gdal = ["masked", "dep:extend", "dep:gdal" ]
+gdal = ["masked", "dep:gdal" ]
 
 [workspace.dependencies]
 once_cell = "1.19.0"
@@ -40,7 +40,6 @@ num-traits = "0.2"
 thiserror = "1"
 paste = "1"
 serde = { version = "1", features = ["derive"], optional = true }
-extend = { version = "1.2", optional = true }
 gdal = { version = "0.16.0", optional = true }
 
 [dev-dependencies]
